Improve rank checking + IK fix for when a stage is locked
Rank checking has been improved and locking a stage now works correctly.
Add a new parameter to the chair script that can be set to ignore setting the player position. Locking the stage will no longer kick speakers or high our of the area (can be easily changed), and any players on stage that are sitting will have be standing up now.